cace size is size_t

+++ benchmark/benchmark.cpp
@@ -24,9 +24,9 @@ int main() {
   size_t rnd_mod = 100;
   std::vector<int> keys = generate_rnd_keys(input_sz, rnd_mod);
 
-  std::map<int, std::map<std::string, int>> cache_sizes_vs_strategy_hits;
+  std::map<size_t, std::map<std::string, int>> cache_sizes_vs_strategy_hits;
   std::cout << "Cache type\tN hits\n";
-  for (int cache_sz = 10; cache_sz < 101; cache_sz += 10) {
+  for (size_t cache_sz = 10; cache_sz < 101; cache_sz += 10) {
     std::map<std::string, int> cache_hits;
     cache::perfect_t<int> perfect_cache(cache_sz, keys);
     cache::LRU_t<int> lru{cache_sz};
